requirements.txt
fastapi
curl_cffi
uvicorn
yfinance
google-genai
python-dotenv
